[refactor] 배치서버 gaspriceavg 부분 리책터리 (gwei 단위로 변경)

+++ batch-server/api/v1/DB/models/blocks.js
@@ -72,7 +72,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       gaspriceavg: {
-        type: DataTypes.DOUBLE(25, 23),
+        type: DataTypes.DOUBLE,
         // type: DataTypes.FLOAT,
         allowNull: true,
       },
